requirements.txt
docopt==0.6.2
hbmqtt==0.9.6
paho-mqtt==2.1.0
passlib==1.7.4
PyYAML==6.0.2
six==1.17.0
transitions==0.9.3
websockets==15.0.1
